* more types added to options   -> GF_OPTION_TYPE_{PATH,XLATOR,PERCENT,TIME,BOOL}.   -> an option is provided to give possible options for type GF_OPTION_TYPE_STR.   -> refer unify.c for more some ideas of options[]
* added string2percent, and string2time functions

+++ scheduler/nufa/src/nufa.h
@@ -45,7 +45,7 @@ struct nufa_struct {
   int32_t local_xl_count; /* Count of the local subvolumes */
 
   uint32_t refresh_interval;
-  uint64_t min_free_disk;
+  uint32_t min_free_disk;
   
   gf_lock_t nufa_lock;
   int32_t child_count;
