+++ scheduler/random/src/random.h
@@ -37,7 +37,7 @@ struct random_sched_struct {
 struct random_struct {
   int32_t child_count;
   uint32_t refresh_interval;
-  uint64_t min_free_disk;
+  uint32_t min_free_disk;
   struct timeval last_stat_entry;
   struct random_sched_struct *array;
   pthread_mutex_t random_mutex;
